Change editor scripts namespace (breaking)
The namespace of editor scripts has been changed to include `.Editor`

This change is intended to remove the possibility of the containing types from being referenced through the `LMirman.RewiredGlyphs` namespace import.

Despite technically being a breaking change, I do not intend for this to increment the major version of the package since the two containing types have absolutely no relevance to the public API.

If the user does have some unexpected use for these types they can still be accessed via the new namespace but they will need to import the additional scope.